Restructure client creation to work for variable google services
